requirements.txt
aiobotocore==2.2.0
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
boto3==1.21.21
botocore==1.24.21
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
et-xmlfile==1.1.0
fastapi==0.78.0
fluent-logger==0.10.0
frozenlist==1.3.0
fsspec==2022.3.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
jmespath==1.0.0
msgpack==1.0.3
multidict==6.0.2
numpy==1.22.4
openpyxl==3.0.10
pandas==1.4.2
pydantic==1.9.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
s3fs==2022.3.0
s3transfer==0.5.2
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
wrapt==1.14.1
XlsxWriter==3.0.3
yarl==1.7.2
